+++ app/src/main/java/org/simple/clinic/home/overdue/OverdueViewEffectHandler.kt
@@ -16,6 +16,7 @@ class OverdueViewEffectHandler(
       OpenSelectDownloadFormatDialog -> uiActions.openSelectDownloadFormatDialog()
       OpenSelectShareFormatDialog -> uiActions.openSelectShareFormatDialog()
       OpenSharingInProgressDialog -> uiActions.openProgressForSharingDialog()
+      OpenOverdueSearch -> uiActions.openOverdueSearch()
     }.exhaustive()
   }
 }