+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Bump Sentry to v6.1.2
 - Consume `OverdueSearchViewEffect` in `OverdueSearchEffectHandler`
 - Implement query to search overdue patients
+- Open overdue search screen when search button is clicked in overdue screen
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-06-20-8300